% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSEIR.R
\name{compute_SEIR}
\alias{compute_SEIR}
\title{Marginal Means Epi Curve}
\usage{
compute_SEIR(
  A_mat,
  I0 = 1,
  nDays = 30,
  nDaysUntilRecovered = 11,
  nDaysSusceptible = 3,
  probTransPerContact = 0.01,
  importProb = 1.05094e-06,
  propPeakToReopen = 0,
  minNumberToReopen = NULL,
  minDayToReopen = Inf,
  gradualReopening = 7,
  nDaysIncreasing = 7,
  secondQuarProb = 0.3,
  quarantineProb = rep(c(0, 0.3, 0.5), c(6, 17, nDays - 17 - 6)),
  quarantineProb_highDegree = quarantineProb,
  highDeg_cutoff = 50,
  plotProgress = TRUE
)
}
\arguments{
\item{A_mat}{(sparse) adjacency matrix}

\item{I0}{integer.  Number of initial infections in the population}

\item{nDays}{integer. Number of days to forecast.}

\item{nDaysUntilRecovered}{integer. Number of days an infective can infect a neighboring susceptible}

\item{probTransPerContact}{numeric (scalar or vector of length nDays). Daily probability 
that a S-I contact will lead to a new transmission event}

\item{importProb}{numeric. Daily probability that a susceptible will import the disease from outside 
of the population}

\item{propPeakToReopen}{numeric. (Adaptive reopening) Quarantining will be relaxed if the number of 
daily new infections reaches propPeakToReopen proportion of the peak.}

\item{minNumberToReopen}{integer. (Adaptive reopening) Quarantining will be relaxed if the number of 
daily new infections is below minNumberToReopen.  If not NULL, minNumberToReopen will override 
propPeakToReopen.}

\item{minDayToReopen}{integer. (Adaptive reopening) Reopening will not occur before minDayToReopen. 
This is important if minNumberToReopen is not NULL, sine the epidemic will be below this threshold in 
its early days.}

\item{gradualReopening}{integer.  (Adaptive reopening) If not NULL, a Gaussian kernel will be used 
to gradually reopen society such that the quarantining probability will reach 90\% of its current 
level after gradualReopening days.}

\item{nDaysIncreasing}{integer.  (Adaptive reopening) Number of days of increasing infections 
before reimplementing social distancing/quarantining to secondQuarProb.}

\item{secondQuarProb}{numeric. (Adaptive reopening) Probability that an individual will be 
quarantined if social distancing needs to be reimplemented.}

\item{quarantineProb}{numeric vector. Vector of length nDays giving the daily probability that an 
individual will be quarantined.}

\item{quarantineProb_highDegree}{numeric vector. Vector of length nDays giving the daily 
probability that an individual above highDeg_cutoff will be quarantined.}

\item{highDeg_cutoff}{integer. The degree cutoff such that anyone with degree higher 
than highDeg_cutoff will follow quarantineProb_highDegree rather than quarantineProb, and if 
adaptive reopening occurs, these high degree individuals will remain at the prespecified 
quarantineProb_highDegree.}

\item{plotProgress}{logical.}
}
\value{
A nDays x 5 matrix giving the cumulative number of susceptibles, cumulative number of infected,
number of daily infectives (not new infections), and a binary vector showing if social distancing 
measures have been relaxed.
}
\description{
computeEpiCurve() computes the expected cumulative number of infections/susceptibles per day.
}
\details{
The effects of nonpharmaceutical interventions on an epidemic can be evaluated using this function.  
It provides the ability to set a schedule of social distancing/quarantining probabilities, of
adaptively reopening society/relaxing social distancing based on a daily hard threshold or by 
proportion of the peak daily number of new infections.  Reopening can occur to the entire population 
or to a subpopulation based on having low number of contacts.
}
