% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNetwork.R
\name{makeNetwork}
\alias{makeNetwork}
\title{Make contact graph}
\usage{
makeNetwork(N, nTries = 10)
}
\arguments{
\item{N}{integer. Size of network/number of nodes.}

\item{nTries}{integer. It sometimes takes a few tries before a network can be made which matches the 
degree distribution.}
}
\value{
igraph object
}
\description{
This function will create a contact network based on observed degree distributions
}
